refactor(hooks): move hook tools into capability
