+++ src/capabilities/mod.rs
@@ -6,6 +6,7 @@
 pub(crate) mod approval;
 pub(crate) mod client_commands;
 pub(crate) mod config;
+pub(crate) mod hooks;
 mod host;
 pub(crate) mod memory;
 pub(crate) mod model_discovery;
@@ -15,6 +16,7 @@ pub(crate) mod your;
 pub(crate) use approval::{APPROVAL_CAPABILITY_ID, ApprovalCapability};
 pub(crate) use client_commands::{CLIENT_COMMANDS_CAPABILITY_ID, ClientCommandsCapability};
 pub(crate) use config::{CONFIG_CAPABILITY_ID, ConfigCapability};
+pub(crate) use hooks::{HOOKS_CAPABILITY_ID, HooksCapability};
 pub(crate) use host::{
     ATTRIBUTION_CAPABILITY_ID, AttributionCapability, CodingBashCapability,
     CodingCliEnvironmentCapability, ENVIRONMENT_CONTEXT_CAPABILITY_ID, SETUP_CAPABILITY_ID,
